0319 백엔드 커밋

+++ src/main/java/com/spring/bittlebittle/review/controller/ReviewController.java
@@ -57,22 +57,6 @@ public class ReviewController {
 
 		return map;
 	}
-
-
-//	@PostMapping
-//	public List<Review> addReview(@PathVariable int bottleNo, 
-//			@ModelAttribute Review review){     
-//		
-//
-//        	review.setBottleNo(bottleNo);
-//        	
-//        	log.debug(review);
-//    		
-//    		List<Review> reviewList = rservice.addReview(review);
-//    		
-//    		return reviewList;
-//		
-//	}
 	
 
 	// 리뷰등록 (확인완료)
@@ -90,8 +74,6 @@ public class ReviewController {
 				.userJwtIdx(refreshTokenIdx)
 				.build())) {
 
-			review.setBottleNo(bottleNo);
-
 			List<Review> reviewList = rservice.addReview(review);
 
 			return reviewList;
@@ -105,25 +87,54 @@ public class ReviewController {
 
 	// 리뷰수정 (확인완료)
 	@PostMapping(value="/set-data")
-	public List<Review> editReveiw(@PathVariable int bottleNo, @ModelAttribute Review review){
+	public List<Review> editReveiw(@PathVariable int bottleNo, @ModelAttribute Review review
+									, HttpServletRequest request){
 																			// reviewNo도 데이터로 보내줘야함
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
 
-		review.setBottleNo(bottleNo);
+			List<Review> reviewList = rservice.editReview(review);
 
-		log.debug(review);
+			return reviewList;
+			
+		} else {
+			
+			List<Review> reviewList = rservice.getReviews(bottleNo);
+			
+			return reviewList;
+		}
+		
 
-		List<Review> reviewList = rservice.editReview(review);
-
-		return reviewList;
+		
 	}
 
 	// 리뷰삭제 (확인완료)
 	@GetMapping(value="/{reviewNo}/deletion")
-	public List<Review> deleteReview(@PathVariable int bottleNo, @PathVariable int reviewNo){
+	public List<Review> deleteReview(@PathVariable int bottleNo, @PathVariable int reviewNo, HttpServletRequest request){
 
-		List<Review> reviewList = rservice.removeReview(bottleNo, reviewNo);
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
 
-		return reviewList;
+			List<Review> reviewList = rservice.removeReview(bottleNo, reviewNo);
+
+			return reviewList;
+			
+		} else {
+			
+			List<Review> reviewList = rservice.getReviews(bottleNo);
+			
+			return reviewList;
+		}
 	}
 
 	// 리뷰댓글조회 (확인완료)
@@ -137,38 +148,88 @@ public class ReviewController {
 
 	// 리뷰댓글작성 (확인완료)
 	@PostMapping(value="/{reviewNo}/replies")
-	public List<Reply> addReply(@PathVariable int reviewNo, @ModelAttribute Reply reply){
+	public List<Reply> addReply(@PathVariable int reviewNo, @ModelAttribute Reply reply, HttpServletRequest request){
 																		// replyContent
-		int userNo = 1;
+		
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
 
-		reply.setReviewNo(reviewNo);
-		reply.setUserNo(userNo);
+			log.debug(reply);
 
-		List<Reply> replyList = rpservice.addReply(reply);
+			List<Reply> replyList = rpservice.addReply(reply);
 
-		return replyList;
+			return replyList;
+			
+		} else {
+			
+			List<Reply> replyList = rpservice.getReplies(reviewNo);
+
+			return replyList;
+		}
+		
+		
+
 	}
 
 	// 리뷰댓글수정 (확인완료)
 	@PostMapping(value="/{reviewNo}/replies/set-data")
-	public List<Reply> updateReply(@PathVariable int reviewNo, @RequestBody Reply reply){
+	public List<Reply> updateReply(@PathVariable int reviewNo, @ModelAttribute Reply reply, HttpServletRequest request){
+																// replyNo, replyContent
 
-		// replyNo, replyContent
+		
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
 
-		reply.setReviewNo(reviewNo);
+			List<Reply> replyList = rpservice.editReply(reply);
 
-		List<Reply> replyList = rpservice.editReply(reply);
+			return replyList;
+			
+		} else {
+			
+			List<Reply> replyList = rpservice.getReplies(reviewNo);
 
-		return replyList;
+			return replyList;
+		}
+		
+
+		
 	}
 
 	// 리뷰댓글삭제 (확인완료)
 	@GetMapping(value="/{reviewNo}/replies/{replyNo}/deletion")
 	public List<Reply> deleteReply(@PathVariable int reviewNo,
-			@PathVariable int replyNo){
-		List<Reply> replyList = rpservice.removeReply(reviewNo, replyNo);
+			@PathVariable int replyNo, HttpServletRequest request){
+		
+		String token = jwtUtil.resolveAccessToken(request);
+		String refreshTokenIdx = jwtUtil.resolveRefreshToken(request);
+		log.debug(token);
+		log.debug(refreshTokenIdx);
+		if (jwtUtil.validateToken(token, UserJwt.builder()
+				.userJwtIdx(refreshTokenIdx)
+				.build())) {
 
-		return replyList;
+
+			List<Reply> replyList = rpservice.removeReply(reviewNo, replyNo);
+
+			return replyList;
+			
+		} else {
+			
+			List<Reply> replyList = rpservice.getReplies(reviewNo);
+
+			return replyList;
+		}
+		
 	}
 
 }
